Exceptions move part1
Problem:
Currently exception classes are scattered all over SDK, this will lead to unecessary duplication and confusion

Analysis:
Removed rest of exceptions from mixins, corrected related files

Tests:
Flake8
