+++ f5/bigip/test/unit/test_mixins.py
@@ -22,11 +22,11 @@ import struct
 
 from f5.bigip.mixins import AsmFileMixin
 from f5.bigip.mixins import CommandExecutionMixin
-from f5.bigip.mixins import EmptyContent
-from f5.bigip.mixins import MissingHttpHeader
 from f5.bigip.mixins import ToDictMixin
 from f5.bigip.mixins import UnsupportedMethod
 from f5.bigip.resource import Resource
+from f5.sdk_exception import EmptyContent
+from f5.sdk_exception import MissingHttpHeader
 
 from requests import HTTPError
